tests/wpt/mozilla/tests/bluetooth/getCharacteristic/get-same-characteristic.https.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/_mozilla/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.heart_rate);
    return window.navigator.bluetooth.requestDevice({
        filters: [{services: [heart_rate.name]}],
        optionalServices: [generic_access.name]
    })
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService(generic_access.name))
    .then(service => {
        return Promise.all([
            service.getCharacteristic(device_name.name),
            service.getCharacteristic(device_name.name)])
        .then(characteristics => {
            for (let i = 1; i < characteristics.length; i++) {
                assert_equals(characteristics[0], characteristics[i],
                    'Should return the same characteristic as the first call.');
            }
        });
    });
}, 'Calls to get the same characteristic should return the same object.');
</script>
